new GDD

+++ Assets/_Project/RicochetTanks/Scripts/Gameplay/Combat/HitResolvedEvent.cs
@@ -4,7 +4,7 @@ namespace RicochetTanks.Gameplay.Combat
 {
     public readonly struct HitResolvedEvent
     {
-        public HitResolvedEvent(TankFacade source, TankFacade target, HitResult result, int damage, int currentHp, int maxHp, ArmorHitInfo armorHit)
+        public HitResolvedEvent(TankFacade source, TankFacade target, HitResult result, float damage, float currentHp, float maxHp, ArmorHitInfo armorHit)
         {
             Source = source;
             Target = target;
@@ -18,9 +18,9 @@ namespace RicochetTanks.Gameplay.Combat
         public TankFacade Source { get; }
         public TankFacade Target { get; }
         public HitResult Result { get; }
-        public int Damage { get; }
-        public int CurrentHp { get; }
-        public int MaxHp { get; }
+        public float Damage { get; }
+        public float CurrentHp { get; }
+        public float MaxHp { get; }
         public ArmorHitInfo ArmorHit { get; }
     }
 }
